Separate unit tests from function tests of  servers.json
